test(graphql): move tests of moved queries
