+++ NineChronicles.Headless.Tests/GraphTypes/StandaloneQueryTest.cs
@@ -529,112 +529,6 @@ namespace NineChronicles.Headless.Tests.GraphTypes
             );
         }
 
-        [Fact]
-        public async Task NextTxNonce()
-        {
-            var userPrivateKey = new PrivateKey();
-            var userAddress = userPrivateKey.ToAddress();
-            var service = MakeMineChroniclesNodeService(userPrivateKey);
-            StandaloneContextFx.NineChroniclesNodeService = service;
-            StandaloneContextFx.BlockChain = service.Swarm?.BlockChain;
-
-            var blockChain = StandaloneContextFx.BlockChain!;
-            var query = $"query {{ transaction {{ nextTxNonce(address: \"{userAddress}\") }} }}";
-            var queryResult = await ExecuteQueryAsync(query);
-            Assert.Equal(
-                new Dictionary<string, object>
-                {
-                    ["transaction"] = new Dictionary<string, object>
-                    {
-                        ["nextTxNonce"] = 0L   
-                    }
-                },
-                queryResult.Data
-            );
-
-            blockChain.MakeTransaction(userPrivateKey, new PolymorphicAction<ActionBase>[] { });
-            queryResult = await ExecuteQueryAsync(query);
-            Assert.Equal(
-                new Dictionary<string, object>
-                {
-                    ["transaction"] = new Dictionary<string, object>
-                    {
-                        ["nextTxNonce"] = 1L
-                    }
-                },
-                queryResult.Data
-            );
-        }
-
-        [Fact]
-        public async Task GetTx()
-        {
-            var userPrivateKey = new PrivateKey();
-            var userAddress = userPrivateKey.ToAddress();
-            var service = MakeMineChroniclesNodeService(userPrivateKey);
-            StandaloneContextFx.NineChroniclesNodeService = service;
-            StandaloneContextFx.BlockChain = service.Swarm!.BlockChain;
-
-            var blockChain = StandaloneContextFx.BlockChain;
-            var queryFormat = @"query {{
-                transaction {{
-                    getTx(txId: ""{0}"") {{
-                        id
-                        nonce
-                        signer
-                        signature
-                        timestamp
-                        updatedAddresses
-                        actions {{
-                            inspection
-                        }}
-                    }}
-                }}
-            }}";
-            var queryResult = await ExecuteQueryAsync(string.Format(queryFormat, new TxId()));
-            Assert.Equal(
-                new Dictionary<string, object?>
-                {
-                    ["transaction"] = new Dictionary<string, object?>
-                    {
-                        ["getTx"] = null
-                    }
-                },
-                queryResult.Data
-            );
-
-            var action = new CreateAvatar2
-            {
-                index = 0,
-                hair = 1,
-                lens = 2,
-                ear = 3,
-                tail = 4,
-                name = "action",
-            };
-            var transaction = blockChain.MakeTransaction(userPrivateKey, new PolymorphicAction<ActionBase>[] { action });
-            blockChain.StageTransaction(transaction);
-            await blockChain.MineBlock(new Address());
-            queryResult = await ExecuteQueryAsync(string.Format(queryFormat, transaction.Id));
-            var tx = queryResult.Data
-                .As<Dictionary<string, object>>()["transaction"]
-                .As<Dictionary<string, object>>()["getTx"]
-                .As<Dictionary<string, object>>();
-
-            Assert.Equal(tx["id"], transaction.Id.ToString());
-            Assert.Equal(tx["nonce"], transaction.Nonce);
-            Assert.Equal(tx["signer"], transaction.Signer.ToString());
-            Assert.Equal(tx["signature"], ByteUtil.Hex(transaction.Signature));
-            Assert.Equal(tx["timestamp"], transaction.Timestamp.ToString());
-            Assert.Equal(tx["updatedAddresses"], transaction.UpdatedAddresses.Select(a => a.ToString()));
-
-            var plainValue = tx["actions"]
-                .As<List<object>>()
-                .First()
-                .As<Dictionary<string, object>>()["inspection"];
-            Assert.Equal(transaction.Actions.First().PlainValue.Inspection, plainValue);
-        }
-
         [Theory]
         [InlineData(null)]
         [InlineData("memo")]
